(chore) extract validator reusable concerns
