+++ app/validators/email_address_validator.rb
@@ -1,23 +1,16 @@
 class EmailAddressValidator < ActiveModel::EachValidator
+  include ValidatorConcerns
   EMAIL_FORMAT = /\A([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})\z/i
 
   def validate_each(record, attribute, value)
-    return record.errors[attribute] << cant_be_blank_message if check_presence? && value.blank?
-    record.errors[attribute] << invalid_error_message(options) if value[EMAIL_FORMAT].nil?
+    return error_message(record, attribute, cant_be_blank_message) if check_presence? && value.blank?
+    error_message(record, attribute, invalid_error_message) if value[EMAIL_FORMAT].nil?
   end
 
   private
 
-  def invalid_error_message(options)
+  def invalid_error_message
     options[:message] || I18n.t('errors.messages.email.invalid')
-  end
-
-  def check_presence?
-    options.key?(:presence) && options[:presence]
-  end
-
-  def cant_be_blank_message
-    I18n.t('errors.messages.blank')
   end
 end
